<feat> Updated Input Files and input file creation
Sometimes an experiment hits the min voltage constraint during simulation. To make it easy to see which experiment hit the constraint, I've added in an exp_num cell to the .mat files which then pass the number to DFN_sim_casadi. sim_data_creator.m puts it into .mat files.